Remove script adding target="_blank" to URL pointing away from the site
I set this up ages ago with the rationale that it would have helped me
retain readers on the site.

How arrogant of me to decided what readers should do. How lazy to try
and keep them on my site with anything other than sharp and useful
writing.

Turns out there's quite a bit of research on the topic, for example
https://www.nngroup.com/articles/new-browser-windows-and-tabs/, and the
general best practice is to NOT open new tabs or windows –let your users
decide where they want to go.

With a tech savvy and opinionated audience like mine, this is true even
more so.